Cleanup layout and dragging code
